shop item styling

+++ BE/Flight2Orbit/Models/Inventory/ResourceDTO.cs
@@ -5,9 +5,9 @@
         public string Title { get; set; }
         public string ColourCode { get; set; }
         public string Type { get; set; }
-        public int Amount { get; set; }
+        public double Amount { get; set; }
 
-        public ResourceDTO(string type, int amount)
+        public ResourceDTO(string type, double amount)
         {
             Type = type;
             Amount = amount;
